requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.15
hnswlib==0.7.0
sentence-transformers==2.2.2
openai==1.10.0
pypdf==4.0.1
docx2txt==0.8
protobuf==3.20.2
pytz==2020.5
torch==2.1.0 